\name{Marble-package}
\alias{Marble-package}
\alias{Marble}
\docType{package}
\title{
\packageTitle{Marble}
}
\description{
\packageDescription{Marble}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Marble}
\packageIndices{Marble}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Marble}

Maintainer: \packageMaintainer{Marble}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
