% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.marble.R
\name{print.marble}
\alias{print.marble}
\title{print a marble object}
\usage{
\method{print}{marble}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{marble object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Print a summary of a marble object
}
\seealso{
\code{\link{marble}}
}
