% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sazareishi.R
\name{sazareishi}
\alias{sazareishi}
\alias{sazareishi_to_df}
\title{Word coloring according to part of speech}
\usage{
sazareishi(x, elem = "text", ...)

sazareishi_to_df(df, token = token, pos = pos, ...)
}
\arguments{
\item{x}{character vector}

\item{elem}{path to emphatic's hl_vec(elem = ...). "text" or "fill".}

\item{...}{
  Arguments passed on to \code{\link[=sazareishi]{sazareishi}}
  \describe{
    \item{\code{}}{}
  }}

\item{df}{data.frame}

\item{token}{name for token variable}

\item{pos}{name for part of speech variable}
}
\description{
Word coloring according to part of speech
}
